feat(niuu): extract shared services into independent plugin with own API app (NIU-412)
Move RepoService, git provider registry, GitHub/GitLab adapters, shared
models, and git port interfaces from volundr to niuu module. Create
NiuuPlugin with its own FastAPI app mounted at /api/v1/niuu. Volundr
re-exports moved types for backward compatibility.
