+++ src/volundr/domain/models.py
@@ -12,11 +12,15 @@ from uuid import UUID, uuid4
 from pydantic import BaseModel, Field
 
 from niuu.domain.models import (  # noqa: F401
+    CIStatus,
     GitProviderType,
     IntegrationConnection,
     IntegrationType,
     Principal,
+    PullRequest,
+    PullRequestStatus,
     RepoInfo,
+    ReviewStatus,
     SecretType,
     StoredCredential,
 )
@@ -608,51 +612,6 @@ class Chronicle(BaseModel):
     )
 
     model_config = {"frozen": False}
-
-
-class PullRequestStatus(StrEnum):
-    """Status of a pull request."""
-
-    OPEN = "open"
-    MERGED = "merged"
-    CLOSED = "closed"
-
-
-class CIStatus(StrEnum):
-    """CI pipeline status."""
-
-    PASSING = "passing"
-    FAILING = "failing"
-    PENDING = "pending"
-    UNKNOWN = "unknown"
-
-
-class ReviewStatus(StrEnum):
-    """Code review status."""
-
-    APPROVED = "approved"
-    CHANGES_REQUESTED = "changes_requested"
-    PENDING = "pending"
-    UNKNOWN = "unknown"
-
-
-@dataclass(frozen=True)
-class PullRequest:
-    """A pull request / merge request from a git provider."""
-
-    number: int
-    title: str
-    url: str
-    repo_url: str
-    provider: GitProviderType
-    source_branch: str
-    target_branch: str
-    status: PullRequestStatus
-    description: str | None = None
-    ci_status: CIStatus | None = None
-    review_status: ReviewStatus | None = None
-    created_at: datetime | None = None
-    updated_at: datetime | None = None
 
 
 @dataclass(frozen=True)
